requirements.txt
Django==3.1.1
mysqlclient==2.0.1
djangorestframework==3.11.1

django-seed

djangorestframework-datatables==0.5.2
